face-similarity-app/python-backend/requirements.txt
# ============================================================================
# Core Web Framework
# ============================================================================
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.1.3             # Flask dependency, explicit for file handling

# ============================================================================
# Database (PostgreSQL via Supabase)
# ============================================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9      # PostgreSQL adapter
python-dotenv==1.0.1        # Environment variable management

# ============================================================================
# AWS S3 Storage
# ============================================================================
boto3==1.34.0               # AWS SDK for S3 operations
botocore==1.34.0            # boto3 core dependency (pinned for stability)

# ============================================================================
# Authentication & Security
# ============================================================================
bcrypt==4.1.2               # Password hashing
pyjwt==2.8.0                # JWT token generation/validation

# ============================================================================
# Face Recognition & Deep Learning
# ============================================================================
deepface==0.0.93            # Face recognition framework (ArcFace, Facenet512)
opencv-python-headless==4.10.0.84  # Computer vision (headless = no GUI deps)
numpy>=1.24,<2.0            # Numerical computing
onnxruntime==1.18.0         # ONNX model runtime
tf-keras==2.20.1            # Keras for TensorFlow models

# ============================================================================
# Vector Search & Similarity
# ============================================================================
faiss-cpu==1.7.4            # Fast similarity search (FAISS)

# ============================================================================
# Testing
# ============================================================================
pytest>=8.0
pytest-cov>=5.0
